+++ src/components/pages/home/Home.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import Banner from "./banner/Banner";
 import { StContainer } from "../../../styles/Container";
-import SectionTitle from "./sectionTitle";
-import HouseContainer from "../../shared/thumbnailContainer/houseContainer";
-import ProductContainer from "../../shared/thumbnailContainer/productContainer";
-import Category from "./category";
+import SectionTitle from "./sectionTitle/SectionTitle";
+import HouseContainer from "../../shared/thumbnailContainer/houseContainer/HouseContainer";
+import ProductContainer from "../../shared/thumbnailContainer/productContainer/ProductContainer";
+import Category from "./category/Category";
 
 const Home = () => {
   const column = { house: "3", product: "4" };
