+++ src/components/pages/home/banner/Banner.jsx
@@ -8,7 +8,7 @@ import {
   banner05,
   banner06,
 } from "../../../../asset/images/banner";
-import Carousel from "../carousel";
+import Carousel from "../carousel/Carousel";
 import { StBanner, StSmallBanner } from "./styled";
 import { Link } from "react-router-dom";
