+++ src/components/shared/thumbnail/Thumbnail.jsx
