+++ src/components/shared/thumbnailContainer/houseContainer/HouseContainer.jsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import { StImgWrap, StThumbnailContainer } from "../styled";
-import Thumbnail from "../../thumbnail";
-import NumberComma from "../../numberComma";
+import Thumbnail from "../../thumbnail/Thumbnail";
+import NumberComma from "../../numberComma/NumberComma";
 import { home_img_01 } from "../../../../asset/images/sample";
 
 function HouseContainer({ articleList, column }) {
